+++ ionoscloud/data_source_nic.go
@@ -1,0 +1,135 @@
+package ionoscloud
+
+import (
+	"context"
+	"fmt"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
+	ionoscloud "github.com/ionos-cloud/sdk-go/v5"
+	"log"
+)
+
+func dataSourceNIC() *schema.Resource {
+	return &schema.Resource{
+		ReadContext: dataSourceNicRead,
+		Schema: map[string]*schema.Schema{
+			"server_id": {
+				Type:         schema.TypeString,
+				Required:     true,
+				ValidateFunc: validation.All(validation.StringIsNotWhiteSpace),
+			},
+			"datacenter_id": {
+				Type:         schema.TypeString,
+				Required:     true,
+				ValidateFunc: validation.All(validation.StringIsNotWhiteSpace),
+			},
+			"id": {
+				Type:     schema.TypeString,
+				Optional: true,
+			},
+			"name": {
+				Type:     schema.TypeString,
+				Optional: true,
+			},
+			"lan": {
+				Type:     schema.TypeInt,
+				Optional: true,
+			},
+			"dhcp": {
+				Type:     schema.TypeBool,
+				Optional: true,
+				Default:  true,
+			},
+			"ips": {
+				Type:     schema.TypeList,
+				Elem:     &schema.Schema{Type: schema.TypeString},
+				Computed: true,
+				Optional: true,
+			},
+			"firewall_active": {
+				Type:     schema.TypeBool,
+				Optional: true,
+			},
+
+			"mac": {
+				Type:     schema.TypeString,
+				Computed: true,
+			},
+		},
+		Timeouts: &resourceDefaultTimeouts,
+	}
+}
+
+func dataSourceNicRead(ctx context.Context, data *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	client := meta.(*ionoscloud.APIClient)
+
+	t, dIdOk := data.GetOk("datacenter_id")
+	st, sIdOk := data.GetOk("server_id")
+	if !dIdOk || !sIdOk {
+		return diag.FromErr(fmt.Errorf("datacenter id and server id must be set"))
+	}
+	var datacenterId, serverId string
+
+	if dIdOk {
+		datacenterId = t.(string)
+	}
+	if sIdOk {
+		serverId = st.(string)
+	}
+	var name string
+	id, idOk := data.GetOk("id")
+
+	t, nameOk := data.GetOk("name")
+	if nameOk {
+		name = t.(string)
+	}
+	var nic ionoscloud.Nic
+	var err error
+	var apiResponse *ionoscloud.APIResponse
+
+	if !idOk && !nameOk {
+		return diag.FromErr(fmt.Errorf("either id, or name must be set"))
+	}
+	if idOk {
+		nic, apiResponse, err = client.NicApi.DatacentersServersNicsFindById(ctx, datacenterId, serverId, id.(string)).Execute()
+		logApiRequestTime(apiResponse)
+		if err != nil {
+			return diag.FromErr(fmt.Errorf("error getting nic with id %s %s", id.(string), err))
+		}
+		if nameOk {
+			if *nic.Properties.Name != name {
+				return diag.FromErr(fmt.Errorf("name of nic (UUID=%s, name=%s) does not match expected name: %s",
+					*nic.Id, *nic.Properties.Name, name))
+			}
+		}
+	} else {
+		nics, apiResponse, err := client.NicApi.DatacentersServersNicsGet(ctx, datacenterId, serverId).Execute()
+		logApiRequestTime(apiResponse)
+
+		if err != nil {
+			return diag.FromErr(fmt.Errorf("an error occured while fetching nics: %s ", err))
+		}
+
+		if nameOk && nics.Items != nil {
+			if len(*nics.Items) > 1 {
+				log.Printf("[WARNING] found multiple nic results for name %s\n", name)
+			}
+			for _, tempNic := range *nics.Items {
+				if tempNic.Properties.Name != nil && *tempNic.Properties.Name == name {
+					nic = tempNic
+					break
+				}
+			}
+		}
+	}
+	if nic.Id == nil {
+		return diag.FromErr(fmt.Errorf("there are no nics that match the search criteria id = %s, name = %s", id, name))
+	}
+
+	if err := NicSetData(data, &nic); err != nil {
+		return diag.FromErr(err)
+	}
+
+	return nil
+}